Report a concretized list of events in addition to a model when counterexamples are found.
