+++ crux-llvm/src/Crux/LLVM/Compile.hs
@@ -238,7 +238,7 @@ makeCounterExamplesLLVM cruxOpts llvmOpts res
     -- skip resource exhausted goals
     NotProvedGoal _ (simErrorReason -> ResourceExhausted{}) _ _ -> return ()
     -- counterexample to non-resource-exhaustion goal
-    NotProvedGoal _ c _ (Just m) ->
+    NotProvedGoal _ c _ (Just (m,_evs)) ->
       do let suff = case plSourceLoc (simErrorLoc c) of
                       SourcePos _ l _ -> show l
                       _               -> "unknown"
